pyproject.toml
[tool.poetry]
authors = ["Tiago Silva <tiago.toya@hotmail.com>"]
description = ""
name = "fast-zero"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = "^1.14.1"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pydantic-settings = "^2.7.1"
pyjwt = "^2.10.1"
python = "3.12.*"
python-multipart = "^0.0.20"
sqlalchemy = "^2.0.38"
tzdata = "^2025.1"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.3"
freezegun = "^1.5.1"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
ruff = "^0.8.3"
taskipy = "^1.14.1"

[tool.taskipy.tasks]
dev = "fastapi dev fast_zero/app.py"
lint = "ruff check . && ruff check . --diff"
lint_fix = "ruff check . --fix && ruff format ."
post_test = "echo 'Executando post_test' && coverage html"
pre_test = "task lint"
test = "pytest -s --cov=fast_zero -vv"

[tool.ruff]
extend-exclude = ['migrations']
line-length = 82

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
addopts = '-p no:warnings'
env_files = ['.env.test']
pythonpath = '.'

[tool.coverage.run]
omit = ['fast_zero/app_v1/*']

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
